backend/socketio_requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
websockets==11.0.3
pydantic==2.4.2
# Socket.IO dependencies
python-socketio>=5.7.0
fastapi-socketio>=0.0.10
